Fix TypeError in sparse matrix construction - ensure integer indices
- Added .astype(int) to row and column indices after concatenation
- scipy.sparse.coo_matrix requires integer indices, not floats
- Defensive fix to handle any upstream float values

+++ src/spectral_fitting_unified_matrix.py
@@ -112,9 +112,9 @@ def build_sparse_matrix_unified(
     Returns:
         Tuple of (sparse_matrix, target_vector, peak_idx_convertor)
     """
-    # Combine matched and unmatched peaks
-    all_row_indices = np.concatenate([matrix_data.row_indices, unmatched_row_indices])
-    all_col_indices = np.concatenate([matrix_data.col_indices, unmatched_col_indices])
+    # Combine matched and unmatched peaks (ensure integer indices)
+    all_row_indices = np.concatenate([matrix_data.row_indices, unmatched_row_indices]).astype(int)
+    all_col_indices = np.concatenate([matrix_data.col_indices, unmatched_col_indices]).astype(int)
     all_values = np.concatenate([matrix_data.values, unmatched_values])
     
     # Rank rows to remove gaps
